chore: migrate grpc-js to instrumentation class to avoid expose internals
