+++ packages/opentelemetry-instrumentation-grpc/src/grpc-js/types.ts
@@ -16,7 +16,7 @@
 
 import type * as grpcJs from '@grpc/grpc-js';
 import type { EventEmitter } from 'events';
-import type { CALL_SPAN_ENDED } from './utils';
+import type { CALL_SPAN_ENDED } from './serverUtils';
 
 /**
  * Server Unary callback type
@@ -52,3 +52,10 @@ export type GrpcClientFunc = ((...args: unknown[]) => GrpcEmitter) & {
   requestStream: boolean;
   responseStream: boolean;
 };
+
+export type ServerRegisterFunction = typeof grpcJs.Server.prototype.register;
+
+export type MakeClientConstructorFunction = typeof grpcJs.makeGenericClientConstructor;
+
+export type { HandleCall } from '@grpc/grpc-js/build/src/server-call';
+export type { PackageDefinition } from '@grpc/grpc-js/build/src/make-client';
