--- packages/opentelemetry-instrumentation-grpc/src/grpc-js/utils.ts
@@ -1,22 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Symbol to include on grpc call if it has already emitted an error event.
- * grpc events that emit 'error' will also emit 'finish' and so only the
- * error event should be processed.
- */
-export const CALL_SPAN_ENDED = Symbol('opentelemetry call span ended');
